refactor: beer review separated
