+++ src/api/review/api.ts
@@ -1,30 +1,10 @@
-import {
-  AllReviewsQueryParams,
-  CreateReviewRequestType,
-} from "@/../hooks/query/useReviewQuery";
-import { BeerReviewsQueryParams } from "@/../typedef/server/beer";
+import { AllReviewsQueryParams } from "@/../hooks/query/useReviewQuery";
 import axios from "axios";
 import {
   LANGUAGE_TYPE,
   ReviewSortEnum,
   UpdatedReviewInfo,
 } from "../../../interface/types";
-
-export const createReviewApi = async (
-  beerId: number,
-  data: CreateReviewRequestType,
-  accessToken: string
-) => {
-  const url = `/api/v1/beers/${beerId}/reviews`;
-  const config = {
-    headers: {
-      Authorization: `Bearer ${accessToken}`,
-    },
-  };
-  const response = await axios.post(url, data, config);
-
-  return response.data;
-};
 
 // 리뷰 1개 delete
 export const deleteReviewApi = async (
@@ -56,27 +36,6 @@ export const fetchAllReviewsApi = async (queryParam: AllReviewsQueryParams) => {
       language,
     },
   });
-  return res.data;
-};
-
-export const fetchBeerReviewsApi = async (
-  queryParams: BeerReviewsQueryParams
-) => {
-  const {
-    beerId,
-    page = 1,
-    size = 10,
-    sort = ReviewSortEnum.RecentlyUpdated,
-  } = queryParams;
-
-  const res = await axios.get(`/api/v1/beers/${beerId}/reviews`, {
-    params: {
-      page,
-      size,
-      sort,
-    },
-  });
-
   return res.data;
 };
